Test for user table in component folder
