+++ admin/src/pages/UserSearch.spec.js
@@ -88,100 +88,6 @@ describe('UserSearch', () => {
       )
     })
 
-    describe('row toggling', () => {
-      it('has 4 users in the table', () => {
-        expect(wrapper.findAll('tbody > tr')).toHaveLength(4)
-      })
-
-      describe('user with email not activated', () => {
-        it('has no details button', () => {
-          expect(
-            wrapper.findAll('tbody > tr').at(3).findAll('td').at(4).find('button').exists(),
-          ).toBeFalsy()
-        })
-
-        it('has a red confirmed button with envelope item', () => {
-          const row = wrapper.findAll('tbody > tr').at(3)
-          expect(row.findAll('td').at(5).find('button').exists()).toBeTruthy()
-          expect(row.findAll('td').at(5).find('button').classes('btn-danger')).toBeTruthy()
-          expect(row.findAll('td').at(5).find('svg').classes('bi-envelope')).toBeTruthy()
-        })
-
-        describe('click on envelope', () => {
-          it('opens the details', async () => {
-            await wrapper
-              .findAll('tbody > tr')
-              .at(3)
-              .findAll('td')
-              .at(5)
-              .find('button')
-              .trigger('click')
-            expect(wrapper.findAll('tbody > tr')).toHaveLength(6)
-            expect(wrapper.findAll('tbody > tr').at(5).find('input').element.value).toBe(
-              'new@user.ch',
-            )
-            expect(wrapper.findAll('tbody > tr').at(5).text()).toContain(
-              'unregister_mail.text_false',
-            )
-          })
-
-          describe('click on envelope again', () => {
-            it('closes the details', async () => {
-              await wrapper
-                .findAll('tbody > tr')
-                .at(3)
-                .findAll('td')
-                .at(5)
-                .find('button')
-                .trigger('click')
-              expect(wrapper.findAll('tbody > tr')).toHaveLength(4)
-            })
-          })
-
-          describe('click on close details', () => {
-            it('closes the details', async () => {
-              await wrapper
-                .findAll('tbody > tr')
-                .at(3)
-                .findAll('td')
-                .at(5)
-                .find('button')
-                .trigger('click')
-              await wrapper.findAll('tbody > tr').at(5).findAll('button').at(1).trigger('click')
-              expect(wrapper.findAll('tbody > tr')).toHaveLength(4)
-            })
-          })
-        })
-      })
-
-      describe('different details', () => {
-        it.skip('shows the creation formular for second user', async () => {
-          await wrapper
-            .findAll('tbody > tr')
-            .at(1)
-            .findAll('td')
-            .at(4)
-            .find('button')
-            .trigger('click')
-          expect(wrapper.findAll('tbody > tr')).toHaveLength(6)
-          expect(
-            wrapper.findAll('tbody > tr').at(3).find('div.component-creation-formular').exists(),
-          ).toBeTruthy()
-        })
-
-        it.skip('shows the transactions for third user', async () => {
-          await wrapper
-            .findAll('tbody > tr')
-            .at(4)
-            .findAll('td')
-            .at(6)
-            .find('button')
-            .trigger('click')
-          expect(wrapper.findAll('tbody > tr')).toHaveLength(6)
-        })
-      })
-    })
-
     describe('unconfirmed emails', () => {
       beforeEach(async () => {
         await wrapper.find('button.btn-block').trigger('click')
